requirements.txt
streamlit
google-generativeai
pypdf
